backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
pillow==9.5.0
numpy==1.24.3
tensorflow==2.15.0 
opencv-python==4.7.0.72
python-dotenv==1.0.0
rasterio==1.3.6
pydantic==1.10.7
h5py==3.10.0